Fix cleanup jobs

+++ app/Espo/Services/AuthLogRecord.php
@@ -67,13 +67,18 @@ class AuthLogRecord extends Record
         }
 
         // delete
-        $toDelete = $this->getEntityManager()->getRepository('AuthLogRecord')
-            ->where(['createdAt<' => (new \DateTime())->modify("-$days days")->format('Y-m-d H:i:s')])
-            ->limit(0, 2000)
-            ->order('createdAt')
-            ->find();
-        foreach ($toDelete as $entity) {
-            $this->getEntityManager()->removeEntity($entity);
+        while (true) {
+            $toDelete = $this->getEntityManager()->getRepository('AuthLogRecord')
+                ->where(['createdAt<' => (new \DateTime())->modify("-$days days")->format('Y-m-d H:i:s')])
+                ->limit(0, 2000)
+                ->order('createdAt')
+                ->find();
+            if (empty($toDelete[0])) {
+                break;
+            }
+            foreach ($toDelete as $entity) {
+                $this->getEntityManager()->removeEntity($entity);
+            }
         }
 
         // delete forever
